Refactor cookie errors
