+++ src/session/store/SessionStoreErrors.ts
