+++ test/session/cookie.test.ts
@@ -1,4 +1,5 @@
 import { expect } from "chai";
+import { extractSignature } from "../../src/utils/CookieUtils";
 import { createSession } from "../utils/SessionGenerator"
 import { generateRandomBytesBase64, generateSessionId, generateSignature } from "../../src/utils/CookieUtils";
 import { Cookie } from "../../src/session/model/Cookie";
@@ -27,7 +28,7 @@ describe("Cookie", () => {
             const signature: string = generateSignature(sessionId, cookieSecret);
 
             const result = Cookie.validateCookieString(cookieSecret, sessionId + signature);
-            expect(result.value).to.deep.eq(sessionId + signature);
+            expect(result.value).to.equal(sessionId + signature);
         });
 
         it("should fail if cookie string is too short", () => {
@@ -37,7 +38,8 @@ describe("Cookie", () => {
             const invalidSignature = "asdkfasd";
 
             [invalidSessionId + validSignature, validSessionId + invalidSignature, invalidSessionId + invalidSignature].forEach(cookie => {
-                expect(() => Cookie.validateCookieString(cookieSecret, cookie)).to.throw();
+                expect(() => Cookie.validateCookieString(cookieSecret, cookie))
+                    .to.throw(`Cookie string is not long enough - it is ${cookie.length} characters long while it should have 55 characters`);
             });
         });
 
@@ -48,7 +50,10 @@ describe("Cookie", () => {
             const sessionId: string = generateSessionId();
             const signature: string = generateSignature(sessionId, cookieCreationSecret);
 
-            expect(() => Cookie.validateCookieString(cookieVerificationSecret, sessionId + signature)).to.throw();
+            const actualSignature = extractSignature(sessionId + signature);
+            const expectedSignature = generateSignature(sessionId + signature, cookieVerificationSecret);
+            expect(() => Cookie.validateCookieString(cookieVerificationSecret, sessionId + signature))
+                .to.throw(`Cookie signature is invalid - it is ${actualSignature} while it should be ${expectedSignature}`);
         });
     });
 });
